requirements.txt
python-dotenv
pandas
mysql-connector-python
selenium
requests
webdriver-manager
python-dotenv
mysql-connector-python
jupyter
branca
jinja2
folium
numpy
gspread
beautifulsoup4
mysql
gspread
beautifulsoup4
mysql
Pillow 
pymysql
python-wordpress-xmlrpc
oauth2client